feat: Add information on Google hacking and OSINT
The code changes include adding information on Google hacking and OSINT to the `info.txt` file in the `content descovy` directory. This information provides examples and descriptions of different filters that can be used for Google hacking, such as `site`, `inurl`, `filetype`, and `intitle`. Additionally, a link to a website for checking website updates has been included.

Note: The recent user commits and repository commits are not relevant for generating the commit message in this case.